requirements.txt
Flask
Flask-SQLAlchemy
supabase
requests
python-dotenv
gunicorn
flask-cors
stripe
